+++ src/mhitm.c
@@ -2478,7 +2478,7 @@ int mdead;
     }
 
     if (mddat->mattk[i].damn > 0 || mddat->mattk[i].damd > 0)
-        basedmg = max(0, d(mddat->mattk[i].damn > 0 ? mddat->mattk[i].damn : mddat->mlevel / 2 + 2, mddat->mattk[i].damd > 0 ? mddat->mattk[i].damd : 6) + (int)mddat->mattk[i].damp);
+        basedmg = max(0, d(mddat->mattk[i].damn > 0 ? mddat->mattk[i].damn : (int)mddat->mlevel / 2 + 2, mddat->mattk[i].damd > 0 ? mddat->mattk[i].damd : 6) + (int)mddat->mattk[i].damp);
     else
         basedmg = max(0, mddat->mattk[i].damp);
